+++ src/rendering/shaders/mandelbulb/sdf/sdf3d.glsl.ts
@@ -2,6 +2,7 @@ export const sdf3dBlock = `
 // ============================================
 // 3D Mandelbulb - Standard with z-axis primary
 // OPT-C5: Defer orbit trap sqrt (minAxisSq)
+// OPT-PREC: mediump for orbit traps (coloring only, not SDF accuracy)
 // ============================================
 
 float sdf3D(vec3 pos, float pwr, float bail, int maxIt, out float trap) {
@@ -14,7 +15,8 @@ float sdf3D(vec3 pos, float pwr, float bail, int maxIt, out float trap) {
     float r = 0.0;
 
     // Orbit traps - OPT-C5: minAxisSq instead of minAxis
-    float minPlane = 1000.0, minAxisSq = 1000000.0, minSphere = 1000.0;
+    // OPT-PREC: mediump sufficient for coloring data
+    mediump float minPlane = 1000.0, minAxisSq = 1000000.0, minSphere = 1000.0;
     int escIt = 0;
 
     // Pre-compute phase offsets outside loop (OPT: saves 2 comparisons per iteration)
